+++ tests/test_self_description.py
@@ -41,6 +41,7 @@ def test_writing_of_dtool_structure_file(tmp_uri_fixture):  # NOQA
         "manifest_relpath": [".dtool", "manifest.json"],
         "overlays_directory": [".dtool", "overlays"],
         "annotations_directory": [".dtool", "annotations"],
+        "tags_directory": [".dtool", "tags"],
         "metadata_fragments_directory": [".dtool", "tmp_fragments"],
         "storage_broker_version": __version__,
     }